+++ frontend/src/components/MainContent.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { StyledDiv } from './MainContent.styles';
+
+const MainContent = ({ children }) => {
+    return <StyledDiv id="main-content">{children}</StyledDiv>;
+};
+
+export default MainContent;
