+++ frontend/src/components/MainContent.styles.js
@@ -1,0 +1,6 @@
+﻿import styled from 'styled-components';
+
+export const StyledDiv = styled.div`
+    overflow: auto;
+    width: 100%;
+`;
