+++ frontend/src/components/buildTable/Table.styles.js
@@ -39,7 +39,7 @@ export const Container = styled.div`
     border-radius: 8px;
     margin: 0 198px 64px 198px;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         margin: 0 108px 24px 108px;
     }
 `;
