+++ frontend/src/components/buttons/Checkbox.jsx
@@ -1,33 +1,12 @@
-// eslint-disable-next-line
-import React, { useState } from 'react';
-import { useQueryParams } from '../hooks/useQuery';
+import React from 'react';
+import { useQueryParams } from '../../hooks/useQuery';
 import { useHistory, useLocation } from 'react-router-dom';
-import styled from 'styled-components';
-
-const CheckboxContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    padding: 20px 40px 20px 0px;
-`;
-
-const StyledCheckbox = styled.input`
-    border: 1px solid #eee;
-    border-radius: 10px;
-    background-color: white;
-    padding: 5px;
-    margin: 5px;
-    cursor: pointer;
-`;
-
-const SelectedCheckbox = styled(StyledCheckbox)`
-    background-color: transparent !important;
-    border: 2px solid #243b53 !important;
-    color: #243b53 !important;
-`;
-
-const Header = styled.h3`
-    padding-left: 7px;
-`;
+import {
+    CheckboxContainer,
+    StyledCheckbox,
+    SelectedCheckbox,
+    Header,
+} from './Checkbox.styles';
 
 const Checkbox = () => {
     const history = useHistory();
