+++ frontend/src/components/dashlist/FailureTable.jsx
@@ -1,41 +1,7 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router';
 import { useStateValue } from '../../contexts/state';
-import styled from 'styled-components';
-
-const TableContainer = styled.div`
-    border-top: solid;
-    border-color: #ddd;
-    max-width: 100%;
-    grid-area: table;
-    max-height: 100%;
-    overflow-y: scroll;
-`;
-
-const StyledTable = styled.table`
-    th:nth-of-type(1) {
-        width: 70%;
-    }
-    th:nth-of-type(2) {
-        width: 20%;
-    }
-    th:nth-of-type(3) {
-        width: 10%;
-    }
-
-    th {
-        border-bottom: 1px solid #ddd;
-        padding-left: 5px;
-        padding-right: 5px;
-        text-align: left;
-    }
-    td {
-        border-bottom: 1px solid #ddd;
-        padding-left: 5px;
-        padding-right: 5px;
-        text-align: left;
-    }
-`;
+import { TableContainer, StyledTable } from './FailureTable.styles';
 
 const DashboardList = () => {
     const { seriesId } = useParams();
