+++ frontend/src/components/dashlist/FlakinessTable.jsx
@@ -1,48 +1,11 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router';
 import { useStateValue } from '../../contexts/state';
-import styled from 'styled-components';
-
-const TableContainer = styled.div`
-    border-top: solid;
-    border-color: #ddd;
-    width: 100%;
-    max-height: 400px;
-    overflow-y: scroll;
-    grid-area: table;
-`;
-
-const StyledTable = styled.table`
-    th:nth-of-type(1) {
-        width: 70%;
-    }
-    th:nth-of-type(2) {
-        width: 20%;
-    }
-    th:nth-of-type(3) {
-        width: 10%;
-    }
-
-    th {
-        border-bottom: 1px solid #ddd;
-        padding-left: 5px;
-        padding-right: 5px;
-        text-align: left;
-    }
-    td {
-        border-bottom: 1px solid #ddd;
-        padding-left: 5px;
-        padding-right: 5px;
-        text-align: left;
-    }
-`;
-
-const HighlightedButton = styled.button`
-    background-color: ${props => props.color};
-    color: var(--gradient-black);
-    border: 1px solid var(--gradient-black);
-    outline: none;
-`;
+import {
+    TableContainer,
+    StyledTable,
+    HighlightedButton,
+} from './FlakinessTable.styles';
 
 const StabilityButton = ({ value, text }) => {
     const [{ stabilityChecker }, dispatch] = useStateValue();
@@ -71,7 +34,6 @@ const DashboardList = () => {
         { testStabilityList, stabilityChecker, amountOfBuilds, offset },
         dispatch,
     ] = useStateValue();
-    console.log('testStabilityList', testStabilityList);
 
     useEffect(() => {
         const url = `/data/series/${seriesId}/most_stable_tests/?builds=${amountOfBuilds}&most=${stabilityChecker}&offset=${offset}`;
