+++ frontend/src/components/overview/Series.styles.js
@@ -6,7 +6,7 @@ export const ParentContainer = styled.div`
     flex-flow: row wrap;
     padding: 0 198px;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         padding: 0 108px;
     }
 `;
