+++ frontend/src/components/parentData/ParentTable.styles.js
@@ -8,7 +8,7 @@ export const Container = styled.div`
     flex-wrap: wrap;
     width: 100%;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         padding: 12px 190px 12px 100px;
     }
 `;
