+++ frontend/src/index.css
@@ -54,24 +54,34 @@ body {
 h1,
 h2,
 h3,
-h4 {
+h4,
+h5 {
     font-family: 'Noto Serif Semibold';
 }
 
 h1 {
     font-size: 40px;
+    font-weight: 700;
 }
 
 h2 {
     font-size: 30px;
+    font-weight: 700;
 }
 
 h3 {
     font-size: 18px;
+    font-weight: 500;
 }
 
 h4 {
-    font-size: 16px;
+    font-size: 20px;
+    font-weight: 500;
+}
+
+h5 {
+    font-size: 14px;
+    font-weight: 600;
 }
 
 svg {
