+++ frontend/src/pages/Build.styles.js
@@ -13,7 +13,7 @@ export const LastRunContainer = styled.div`
 export const TableHeader = styled.h2`
     padding: 40px 198px 0 198px;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         padding: 40px 108px 0 108px;
     }
 `;
