+++ frontend/src/pages/Frontpage.styles.js
@@ -5,7 +5,7 @@ export const FrontPage = styled.main`
     width: 100%;
     padding: 24px 0 24px 108px;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         padding: 12px 0 12px 48px;
     }
 `;
