+++ frontend/src/pages/History.styles.js
@@ -15,7 +15,7 @@ export const FilterContainer = styled(HeritanceContainer)`
     margin: 24px 0;
     padding-left: 198px;
 
-    @media only screen and (max-width: 1280px) {
+    @media only screen and (max-width: 1024px) {
         padding-left: 108px;
         max-width: 800px;
     }
@@ -25,10 +25,10 @@ export const ParentContainer = styled(HeritanceContainer)`
     padding: 24px 0;
     flex-direction: column;
 
-    h3 {
+    h4 {
         padding-left: 198px;
 
-        @media only screen and (max-width: 1280px) {
+        @media only screen and (max-width: 1024px) {
             padding-left: 108px;
         }
     }
