+++ frontend/src/pages/Team.jsx
@@ -1,26 +1,11 @@
 import React, { useEffect } from 'react';
 import { useStateValue } from '../contexts/state';
-import Card from '../components/Card';
+import Card from '../components/team/Card';
 import { useParams } from 'react-router';
 import SelectedTeam from '../components/team/SelectedTeam';
 import Loading from '../components/Loading';
 import { useTranslation } from 'react-i18next';
-import styled from 'styled-components';
-
-const TeamHeader = styled.h2`
-    padding: 24px 0px 24px 198px;
-
-    @media only screen and (max-width: 1280px) {
-        padding: 12px 0 12px 108px;
-    }
-`;
-
-const TeamContainer = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    padding: 40px 108px;
-    background: var(--hermanni-grey-lighter);
-`;
+import { TeamHeader, TeamContainer } from './Team.styles';
 
 const Team = () => {
     const [t] = useTranslation(['team']);
